Adds Promise support for composition
This functionality ended up not needing a separate composeP or pipeP
command like Ramda has. There was a good suggestion on the Ramda project
about using a new `R.then()` helper in their compose and deprecating
composeP/pipeP entirely.

However, for us, once we venture off into Promise land, we need to be
able to resolve the promise to get the chained context. Because we are
checking whether the context itself is now a Promise (as well as any
return values from functions), we know whether we need to be chaining
or not. This enables us to handle all the heavy lifting on behalf of
the user so they don't have to worry about whether the functions they
are composing together return Promises or not.

Resolves: #2